+++ client/src/components/Footer/Footer.scss
@@ -2,7 +2,7 @@
   cursor: pointer;
   font-size: 2.2rem;
 }
-.background-header {
+.background-footer {
   // background-color: #faaf08;
   // background-color: #ffdb5c;
   // background-color: #f77604;
