+++ client/src/components/Footer/Footer.tsx
@@ -1,0 +1,14 @@
+import "bootstrap/dist/css/bootstrap.min.css";
+import './Footer.scss';
+import React from 'react';
+
+
+function Footer() {
+  return (
+    <footer  className="w-100 p-3 background-header text-dark d-flex justify-content-between">
+      <div className="ms-4 h4 d-flex align-items-md-center color-header">Footer</div>
+    </footer>
+  );
+}
+
+export default Footer;
