+++ client/src/components/Page/Page.scss
@@ -1,3 +1,6 @@
-.page {
-  min-height: 700px;
+// .page {
+//   min-height: 700px;
+// }
+.main {
+  min-height: 600px;
 }
